+++ shoop/core/migrations/0024_product_archived.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('shoop', '0023_add_shipment_identifier'),
+    ]
+
+    operations = [
+        migrations.AddField(
+            model_name='product',
+            name='archived',
+            field=models.BooleanField(default=False, verbose_name='archived', db_index=True, editable=False),
+        ),
+    ]